Reduction of types  branch v1-1-beta

+++ OdbcStatement.cpp
@@ -1053,7 +1053,7 @@ bool OdbcStatement::setValue(DescRecord *record, int column)
 			const char *string = RESULTS (getString (column));
 
 			int dataRemaining = strlen(string) - dataOffset;
-			int len = MIN(dataRemaining, (long)MAX(0, bufferLength-1));
+			int len = MIN(dataRemaining, (long)MAX(0, (long)bufferLength-1));
 			 
 			//Added by PR. If len is negative we get an AV
 			//Added by NOMEY. and empty strings have len = 0
@@ -1255,7 +1255,7 @@ RETCODE OdbcStatement::setValue(Binding * binding, int column)
 			}
 			else
 			{
-				int len = MIN(dataRemaining, MAX(0, binding->bufferLength-1));
+				int len = MIN(dataRemaining, MAX(0, (long)binding->bufferLength-1));
 				 
 				if ( len > 0 ) 
 				{
